Chat component refactor, minor styles correction
