+++ src/components/Chat/Chat.css
@@ -1,5 +1,5 @@
 .chat {
-    padding-top: 1.375em;
+    padding: 1.375em 0;
 }
 
 .chat .container {
@@ -33,6 +33,7 @@
     justify-content: center;
     align-items: center;
     border-radius: 100%;
+    background-color: var(--neutral-color);
     position: relative;
 }
 
@@ -41,10 +42,7 @@
     height: 100%;
 }
 
-.chat .user .user-info .profile-image .not-found {
-    background-color: var(--neutral-color);
-    padding: 25% 25%;
-    border-radius: 100%;
+.chat .user .user-info .profile-image .not-found{
     user-select: none;
 }
 
@@ -176,22 +174,50 @@
     border-radius: 100%;
 }
 
+.user-status-dot{
+    display:block;
+    width: 10px;
+    height: 10px;
+    position: absolute;
+    background-color: var(--status-color);
+    bottom: 0;
+    left: 70%;
+    border-radius: 100%;
+    border: 2px solid #fff;
+}
+
+.online{
+    --status-color: green;
+}
+
+.offline{
+    --status-color: rgb(122, 2, 2);
+}
+
 .chatMessages .messagesContainer{
-    overflow-y: scroll;
-    height: 51vh;
-    margin-bottom: 1em;
+    padding: 4em 0 2em 0;
+    z-index: 0;
 }
 
 .chatMessages .chatHeader{
+    box-sizing: border-box;
+    background-color: #fff;
+    padding: 1.3em 7%;
+    position: fixed;
+    top: 6.2em;
+    width: 100%;
+    left: 0;
     display: flex;
     justify-content: space-between;
-    padding-bottom: 1.3em;
+    align-items: center;
     border-bottom: 2px solid var(--neutral-color);
     margin-bottom: 1.3em;
+    z-index:1;
 }
 
 .chatMessages .chatHeader .closeChatBtn{
     cursor: pointer;
+    font-size: 24px;
 }
 
 .chatMessages .chatHeader .user-info{
@@ -215,15 +241,16 @@
 
 .chatMessages .messagesContainer .message{
     position: relative;
+    left: -0.3em;
     background-color: var(--neutral-color);
     padding: 0.7em;
     border-radius: 10px;
-    width: 268px;
+    width: 14rem;
     margin-bottom: 2em;
 }
 
 .chatMessages .messagesContainer .own{
-    left: 1.7em;
+    left: 5%;
     background-color: #33373E;
 }
 
@@ -238,13 +265,20 @@
 .chatMessages .messagesContainer .message .date{
     position: absolute;
     font-size: 7px;
-    left: 32em;
+    left: 70%;
     top: 110%;
 }
 
 .messageInputContainer{
+    background-color: #fff;
     display: flex;
+    box-sizing: border-box;
+    position: fixed;
     align-items: center;
+    left: 0;
+    bottom: 4em;
+    padding: 1em 5%;
+    width: 100%;
     gap: 1em;
 }
